pjt09/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.7
django-cleanup==5.2.0
django-imagekit==4.1.0
django-jquery==3.1.0
fonttools==4.28.1
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
pilkit==2.0
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.1
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.0.0
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wordcloud @ file:///C:/Users/grace/Desktop/ssafy6/4.pjt/wordcloud-1.8.1-cp39-cp39-win_amd64.whl
